# openapi-spec.json
{"openapi":"3.0.1","info":{"title":"FleetGuard360 API","description":"Documentación de la API para FleetGuard360","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/reservations":{"post":{"tags":["reservation-controller"],"operationId":"createReservation","parameters":[{"name":"FleetGuard-Auth-Token","in":"cookie","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponseDTO"}}}}}}},"/api/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignupDTO"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDTO"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDTO"}}}}}}},"/api/routes":{"get":{"tags":["route-controller"],"operationId":"getRoutes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RouteDTO"}}}}}}}},"/api/auth/validate":{"get":{"tags":["auth-controller"],"operationId":"validateToken","parameters":[{"name":"FleetGuard-Auth-Token","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDTO"}}}}}}},"/api/auth/signup/complete":{"get":{"tags":["auth-controller"],"operationId":"completeSignup","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDTO"}}}}}}},"/api/auth/login/complete":{"get":{"tags":["auth-controller"],"operationId":"completeLogin","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageDTO"}}}}}}}},"components":{"schemas":{"ReservationRequestDTO":{"type":"object","properties":{"routeId":{"type":"string"}}},"ReservationResponseDTO":{"type":"object","properties":{"confirmationNumber":{"type":"string"},"message":{"type":"string"},"error":{"type":"string"}}},"SignupDTO":{"type":"object","properties":{"email":{"type":"string"},"phoneNumber":{"type":"string"}}},"MessageDTO":{"type":"object","properties":{"message":{"type":"string"},"error":{"type":"string"}}},"LoginDTO":{"type":"object","properties":{"email":{"type":"string"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"RouteDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"firstStopName":{"type":"string"},"lastStopName":{"type":"string"},"startTime":{"$ref":"#/components/schemas/LocalTime"},"endTime":{"$ref":"#/components/schemas/LocalTime"},"numberOfStops":{"type":"integer","format":"int32"},"price":{"type":"integer","format":"int32"}}}}}}