# src/main/java/com/FleetGuard360F3/DTO/SignupDTO.java
package com.FleetGuard360F3.DTO;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class SignupDTO {
    private String email;
    private String phoneNumber;
}